Centralize jobs.yml parsing inside one script
We needed a script that could output a job's build target, which is in
jobs.yml.  Rather than writing a third script that parses that file,
this commit centralizes that job and makes the two existing scripts
simply call it.

get-promotion-tags is a tiny shell wrapper that can likely be removed
after a simple search-and-replace in our calls to it.